client/requirements.txt
depthai==2.30.0.0
numpy==2.3.0
opencv-python==4.11.0.86
pyserial==3.5
websockets==15.0.1
